+++ build-binary-from-source.sh
@@ -1,4 +1,0 @@
-cd ..
-R CMD INSTALL --build --no-multiarch ooacquire_0.4.4-2.tar.gz
-cd ./ooacquire
-